Move appearance enum to BLEProtocol
There were two different enums (in GapAdvertisingData.h and blecommon.h) that
defined the appearance values for the BLE Appearance characteristic. This code
is removed and a single copy of the enum is included in BLEProcotol.h to
continue the effort of putting general information defined by the BLE spec in
a single file.