Add support for directed advertising mode
The old code was incorrect in stating that the interval must be 0 in directed mode. It *can* be 0, and in that case 'high duty'
mode is used where the avertising interval is at most 3.75 ms.

This commit also includes a very minor patch, in that the minimum advertising interval for ADV_SCANNABLE_UNDIRECTED should be 100 ms.